nofomo/requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohappyeyeballs==2.4.4
aiohttp==3.8.6
aiomysql==0.1.1
aiosignal==1.3.2
alembic==1.10.4
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.95.2
frozenlist==1.5.0
gunicorn==20.1.0
h11==0.14.0
idna==3.10
magic-filter==1.0.12
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.1
pydantic==1.10.21
pydantic_core==2.27.2
PyJWT==2.6.0
PyMySQL==1.1.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==1.4.54
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.20.0
yarl==1.18.3
